leaky_bucket_api/requirements.txt
Django
django-cors-headers
django-mysql
djangorestframework
djangorestframework-jwt
gunicorn
ipython
mysqlclient
PyJWT
pymysql
python-dotenv
pytz
sqlparse
typing_extensions
tzdata
uuid
